Polish up Makefile
manpages should depend on their input -- otherwise Make considers them
up to date as soon as the file by the same name exists. This sort of
works as they are indeed filenames, but they never get updated when
their source files do.

Install all manpages automatically and autodetect their man section.
Adding new manpages should be as simple as creating the right source
file in docs/ and detecting it via wildcard.

The default PREFIX shall always be /usr/local -- if someone wants to
install this outside of a package manager, the fact that it is meant to
interact with distro packages is not justification for installing
untracked files directly to /usr/{bin,share} by default.

Add clean target to provide some way of cleaning generated files which
doesn't rely on assuming a git checkout for `git clean`.

Don't squelch Make's status messages for commands run in the install
target. It is inconsistent with what is done for every other target.