[repro] Rework PKGBUILD fetching
Some PKGBUILDs use _pkgver and so on, so we are unable to count on them being
"correct" when fetching PKGBUILDs. We also never implemented support for epoch.
Another problem is that we can't go over the commit history and look at `updpkg:
...` values as they might not be there.

Instead we traverse the history and find the correct PKGBUILD by checksumming it
against the one in the PKGBUILD.

Signed-off-by: Morten Linderud <morten@linderud.pw>