repro: Rewrite chroot handling
Downgrading the chroot initialized from bootstrap became complicated. It
does things we normally don't have to deal with when using pacstrap,
such as initialized users and groups from tmpfiles/sysusers and systemd
pre-installed.

Rework so the bootstrap container, called "root", is a pivot container.
When we build a package we mount an empty directory, initialize the
keyring, then dump packages into the empty chroot with an overloaded
gpgdirectory.

This is twice as fast as the currrent method of doing two installations
of fetched packages. It also reduces locking complexity and having to
care about overlayfs.

Signed-off-by: Morten Linderud <morten@linderud.pw>