repro: Set faked-system-time in gpg.conf
Old packages would previously fail to build if they depended on a
package that was signed with an expired key, from the perspective of
the keyring they were signed with. We can fix this by setting
faked-system-time to control in gpg.conf to the package build time.

This reveals a further issue that the keyring master key will be
generated, and sign the archlinux master keys, at the time repro
builds that keyring, which may be in the future relative to the build
time of packages that depend on it. As gpg ignores signatures and keys
which appear to be from the future, this will also break the build. We
can fix this by also setting faked-system-time to the *keyring*
package build time when setting up the keyring.

Finally, to ease transition, we discard and rebuild any keyrings in
the cache that weren't built in this way. This should allow recovery
from this issue without manual intervention.

Fixes #87