Add build-package locking
EPHEMERAL in systemd-nspawn uses implicit overlayfs mounts to provide
the container. If the root container is being updated or files are in
the lower directory disappear the results are unspecified and might
cause weird behaviour.

Thus we acquire read locks on the build container to ensure nothing gets
a write lock. The code is weird because the locking mechanism here is
implicit as opposed to explicit in the top level of cmd_check.