Make slock shared and add "keyring" locking
Currently the "shared" lock - slock - is exclusive. Fix that alongside
where it us used. In particular:

We need an exclusive lock for the writer and a shared lock for the
reader. Ensure the write lock is before the test/mkdir or otherwise we
can end within the readlock, with an empty directory - doh.

Co-authored-by: Morten Linderud <morten@linderud.pw>
Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>
---
[Emil]
 - Commit message
 - Split from a larger patch
 - Lock before the test/mkdir & document why
 - Drop if check for of blocking lock
 - Remove duplicate mkdir