Stop early on first archive extension that exists
When a .zst package is downloaded from the archive, there's no point to
also check for the same package with a .xz extension. Doing so results
in huge delays because archive.archlinux.org redirects to archive.org
when a requested file doesn't exist, and archive.org lookups are slow.

Testing on thunar-4.16.11-2, downloads are reduced from 10m36s to 1m39s.