fix: use correct makepkg.conf location without silently failing
There was a typo when implementing the new devtools makepkg.conf
location. Unfortunately the code failed silently which made it not be
detected until now.
Fix this by using the correct location as well as switch to a simple
fall-through approach so we can easily fail instead of silently and
invisibly doing the wrong thing leading to false-negative unreproducible
packages.

Fixes 63ce7514966f28299cfc8e2508e7045e610ee135